Structure HTML, set banner, create form
For the logo to appear solid while the background was opaque, I had to
use a pseudo selector to set a background that would not be linked to
logo div.

The form border was giving me trouble because there was still a black
line on the text field when an input was focused. Had to remove border
that is preset on input fields.